Ignore warnings in run-tests action
stestr may be failing on CI because of DeprecationWarnings raised by
tarfile on python3.12. Juliacall will download Julia as a tar file on
Ubuntu runners, extract the file, and install Julia. Other platform
runners get different files. With python3.12, tarfile raised a
DeprecationWarning about a change in the interface. It appears that
stestr will fail on test discovery if a warning is raised. Therefore,
only the Python3.12 tests appeared to fail on Linux devices where Julia
was not already installed. Locally, an installed version of Julia meant
that tests passed as juliacall didn't have to install it from a Tar
file.