Fix Julia installation raising warning during stestr
It appears stestr does not pass PYTHONWARNINGS to subunit calls to
tests. This means that warnings are still raised by tests, which fail
stestr. This is useful if we're testing for warnings; however,
installing Julia appears to raise some deprecation warnings which we
do not want to fail on. Here we import from *.pauli_propagation to force
Julia installation and print a message if this succeeds. Failure
indicates that Julia installation failed too.